requirements.txt
uiautomation
pillow
pyzbar
aiohttp
asyncio